+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - `SameSite` flag in cookies.
 - Option to register http method in route callbacks in `Server.registerRouteCallback`.
 - Route callback can be registered as file in `rsconfig.json`.
+- Using promises in route callbacks.
 ### Breaking changes
 - Removed `Button` component.
 - Removed `res.render` function.